moved css files to designated folder
