+++ client/src/App.js
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import './App.css';
+import './styles/App.css';
 
 function App() {
   return (
