+++ client/src/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import './normalize.css';
+import './styles/normalize.css';
 import App from './App';
 import * as serviceWorker from './serviceWorker';
